+++ lib/src/validators/string.dart
@@ -3,7 +3,7 @@ import 'package:jaguar_validate/jaguar_validate.dart';
 final _emailRegexp = RegExp(
     r"^((([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+(\.([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+)*)|((\x22)((((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(([\x01-\x08\x0b\x0c\x0e-\x1f\x7f]|\x21|[\x23-\x5b]|[\x5d-\x7e]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(\\([\x01-\x09\x0b\x0c\x0d-\x7f]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]))))*(((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(\x22)))@((([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))\.)+(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))$");
 
-Validator<String> isNotEmpty({bool trim = false, err = 'should not be empty'}) {
+ValidationRule<String> isNotEmpty({bool trim = false, err = 'should not be empty'}) {
   return (String value) {
     if (value == null) return null;
 
@@ -15,7 +15,7 @@ Validator<String> isNotEmpty({bool trim = false, err = 'should not be empty'}) {
   };
 }
 
-Validator<String> isEmailAddress({err = 'not a valid email address'}) {
+ValidationRule<String> isEmailAddress({err = 'not a valid email address'}) {
   return (String value) {
     if (value == null) return null;
 
@@ -27,7 +27,7 @@ Validator<String> isEmailAddress({err = 'not a valid email address'}) {
 String _hasLengthErrFunc(int expected, int actual) =>
     'must have length $expected';
 
-Validator<String> hasLength(int length, {err = _hasLengthErrFunc}) {
+ValidationRule<String> hasLength(int length, {err = _hasLengthErrFunc}) {
   return (String value) {
     if (value == null) return null;
 
@@ -40,7 +40,7 @@ Validator<String> hasLength(int length, {err = _hasLengthErrFunc}) {
 String _hasMinLengthErrFunction(int min, int actual) =>
     'should be at least $min characters long';
 
-Validator<String> hasMinLength(int length, {err = _hasMinLengthErrFunction}) {
+ValidationRule<String> hasMinLength(int length, {err = _hasMinLengthErrFunction}) {
   return (String value) {
     if (value == null) return null;
 
@@ -53,7 +53,7 @@ Validator<String> hasMinLength(int length, {err = _hasMinLengthErrFunction}) {
 String _hasMaxLengthErrFunction(int max, int actual) =>
     'cannot be longer than $max characters';
 
-Validator<String> hasMaxLength(int length, {err = _hasMaxLengthErrFunction}) {
+ValidationRule<String> hasMaxLength(int length, {err = _hasMaxLengthErrFunction}) {
   return (String value) {
     if (value == null) return null;
 
@@ -66,7 +66,7 @@ Validator<String> hasMaxLength(int length, {err = _hasMaxLengthErrFunction}) {
 String _hasLengthInRangeErrFunction(int min, int max, int actual) =>
     'length should be in range [$min, $max]';
 
-Validator<String> hasLengthInRange(int min, int max,
+ValidationRule<String> hasLengthInRange(int min, int max,
     {err = _hasLengthInRangeErrFunction}) {
   return (String value) {
     if (value == null) return null;
@@ -81,7 +81,7 @@ Validator<String> hasLengthInRange(int min, int max,
 
 final _anAlphabetRegexp = RegExp(r'[a-zA-Z]');
 
-Validator<String> hasAnAlphabet(
+ValidationRule<String> hasAnAlphabet(
     {err = 'should contain at least one alphabet'}) {
   return (String value) {
     if (value == null) return null;
@@ -95,7 +95,7 @@ Validator<String> hasAnAlphabet(
 
 final _aSpecialCharacterRegexp = RegExp(r'[^A-Za-z0-9 ]');
 
-Validator<String> hasASpecialChar(
+ValidationRule<String> hasASpecialChar(
     {err = 'should contain at least one special character'}) {
   return (String value) {
     if (value == null) return null;
@@ -109,7 +109,7 @@ Validator<String> hasASpecialChar(
 
 final _isAlphaRegexp = RegExp(r'^[a-zA-Z]+$');
 
-Validator<String> isAlpha({err = 'should contain only alphabets'}) {
+ValidationRule<String> isAlpha({err = 'should contain only alphabets'}) {
   return (String value) {
     if (value == null) return null;
 
@@ -122,7 +122,7 @@ Validator<String> isAlpha({err = 'should contain only alphabets'}) {
 
 final _alphaNumericRegexp = RegExp(r'^[a-zA-Z0-9]+$');
 
-Validator<String> isAlphaNumeric(
+ValidationRule<String> isAlphaNumeric(
     {err = 'should contain only alphabets and numbers'}) {
   return (String value) {
     if (value == null) return null;
@@ -136,7 +136,7 @@ Validator<String> isAlphaNumeric(
 
 final _numericRegexp = RegExp(r'^[0-9]+$');
 
-Validator<String> isNumeric({err = 'should contain only numbers'}) {
+ValidationRule<String> isNumeric({err = 'should contain only numbers'}) {
   return (String value) {
     if (value == null) return null;
 
@@ -149,7 +149,7 @@ Validator<String> isNumeric({err = 'should contain only numbers'}) {
 
 final _aNumberRegexp = RegExp(r'[0-9]');
 
-Validator<String> hasANumber([err = 'should contain at least one number']) {
+ValidationRule<String> hasANumber([err = 'should contain at least one number']) {
   return (String value) {
     if (value == null) return null;
 
@@ -163,7 +163,7 @@ Validator<String> hasANumber([err = 'should contain at least one number']) {
 String _maxRepeatedCharactersErrFunction(int allowedRepeats) =>
     'should not have same character repeating consecutively more than ${allowedRepeats} times';
 
-Validator<String> maxRepeatedCharacters(
+ValidationRule<String> maxRepeatedCharacters(
     {int allowedRepeats = 3, err = _maxRepeatedCharactersErrFunction}) {
   final regexp = RegExp(r'(.)\1{' + (allowedRepeats + 1).toString() + r',}');
   return (String value) {
@@ -180,7 +180,7 @@ Validator<String> maxRepeatedCharacters(
 String _matchesRegExpErrFunction(RegExp regexp) =>
     'should match pattern "${regexp.pattern}"';
 
-Validator<String> matchesRegExp(RegExp regexp,
+ValidationRule<String> matchesRegExp(RegExp regexp,
     {err = _matchesRegExpErrFunction}) {
   return (String value) {
     if (value == null) return null;