+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+# 2.0.0
+
++ Total rewrite
++ `validateValue` to validate individual field with validators
++ `ObjectErrors` for object and array validation
++ `ValidationError` to support client side translation of error messages
+
 ## 1.0.0
 
 - Basic `String`, `int`, `double` and `num` validators